+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Changelog for `cleff`
 
-## Unreleased
+## 0.2.1.0 (2022-02-13)
 
 ### Added
 
-- Convenience lifted instances of `Bounded`, `Num`, `Fractional`, `Floating` and `IsString` for `Eff`
+- Lifted convenience instances of `Bounded`, `Num`, `Fractional`, `Floating` and `IsString` for `Eff`
 - `MonadZip` instance from the `MonadComprehensions` extension for `Eff`
 - `freshEnumToState` and `runFreshAtomicCounter` for `Fresh`
 - `inputToReader`, `mapInput` and `bindInput` for `Input`